requirements.txt
fastapi
pydantic
sqlmodel
python-jose
cryptography
sqlalchemy
pymysql
dotenv
pydantic[email]
pymysql
psycopg2

